When the archive is exported to Mbox, any quoted messages appearing in the email body are included.
A bug in the Mstor library resulted in sometimes messages being ignored
when importing an Mbox file. Updated to Mstor version 1.0.2 where the
issue has been fixed.

Exported Mbox files contained all attachments in Base64 encoding. Now
the encoding is read during the Mbox import and then used in the export.

In exported Mbox files encapsulated messages (RFC822) used to be added
to the email body. Now they are kept as attachments.

The HTML part of a message used to be discared during import. Now the
HTML part is stored and used in the export to Mbox.

All headers present in the imported emails are now included when
exporting to Mbox.